(SERVER-2265) Refactor: move cli parsing
