+++ lib/puppetserver/ca/action/generate.rb
@@ -2,7 +2,7 @@ require 'optparse'
 require 'openssl'
 require 'puppetserver/ca/utils/file_system'
 require 'puppetserver/ca/host'
-require 'puppetserver/ca/utils'
+require 'puppetserver/ca/utils/cli_parsing'
 require 'puppetserver/utils/signing_digest'
 
 module Puppetserver
