+++ lib/puppetserver/ca/action/import.rb
@@ -2,7 +2,7 @@ require 'optparse'
 require 'puppetserver/ca/utils/file_system'
 require 'puppetserver/ca/x509_loader'
 require 'puppetserver/ca/puppet_config'
-require 'puppetserver/ca/utils'
+require 'puppetserver/ca/utils/cli_parsing'
 
 module Puppetserver
   module Ca
