+++ lib/puppetserver/ca/action/revoke.rb
@@ -1,4 +1,4 @@
-require 'puppetserver/ca/utils'
+require 'puppetserver/ca/utils/cli_parsing'
 require 'puppetserver/ca/utils/http_client'
 require 'puppetserver/ca/utils/file_system'
 require 'puppetserver/ca/puppet_config'
