+++ lib/puppetserver/ca/utils/cli_parsing.rb
